requirements.txt
asgiref==3.5.0
Django==4.0.3
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2
python-dotenv==0.19.2
django-elasticsearch-dsl==7.2.2
elasticsearch==7.17.1
elasticsearch-dsl==7.4.0